+++ recommender/engines/autoencoders/training/model_evaluation_step.py
@@ -1,12 +1,18 @@
-# pylint: disable=line-too-long,
+# pylint: disable=line-too-long, too-many-locals
 """Autoencoder Model Evaluation Step."""
 
 from typing import Tuple
+import time
 import torch
 from torch.utils.data import DataLoader
-
 from recommender.engines.base.base_steps import ModelEvaluationStep
-from recommender.engines.constants import DEVICE
+from recommender.engines.constants import (
+    DEVICE,
+    METRICS,
+    LOSS,
+    ACCURACY,
+    VERBOSE,
+)
 from recommender.engines.autoencoders.training.model_training_step import (
     MODEL,
     DATASET,
@@ -22,10 +28,12 @@ from recommender.engines.autoencoders.training.data_extraction_step import (
     USERS,
     SERVICES,
 )
+from recommender.engines.metadata_creators import accuracy_function
+from logger_config import get_logger
+
+logger = get_logger(__name__)
 
 BATCH_SIZE = "batch_size"
-METRICS = "metrics"
-LOSS = "loss"
 
 
 class AEModelEvaluationStep(ModelEvaluationStep):
@@ -35,6 +43,7 @@ class AEModelEvaluationStep(ModelEvaluationStep):
         super().__init__(pipeline_config)
         self.device = self.resolve_constant(DEVICE, torch.device("cpu"))
         self.batch_size = self.resolve_constant(BATCH_SIZE, 128)
+        self.verbose = self.resolve_constant(VERBOSE, True)
 
     def __call__(self, data=None) -> Tuple[object, dict]:
 
@@ -43,16 +52,29 @@ class AEModelEvaluationStep(ModelEvaluationStep):
             SERVICES: {TRAIN: {}, VALID: {}, TEST: {}},
         }
 
+        start_evaluation = time.time()
+
         for collection_name, datasets in data.items():
             model = datasets[MODEL]
             for split, dataset in datasets[DATASET].items():
                 dataloader = DataLoader(
                     dataset, batch_size=self.batch_size, shuffle=True
                 )
-                loss = evaluate_autoencoder(
-                    model, dataloader, autoencoder_loss_function, self.device
+                loss, acc = evaluate_autoencoder(
+                    model,
+                    dataloader,
+                    autoencoder_loss_function,
+                    accuracy_function,
+                    self.device,
                 )
-                metrics[collection_name][split] = loss
+                metrics[collection_name][split][LOSS] = loss
+                metrics[collection_name][split][ACCURACY] = acc
+
+        if self.verbose:
+            logger.info(
+                "Evaluation total duration: %ss",
+                round(time.time() - start_evaluation, 3),
+            )
 
         details = {METRICS: metrics}
         data[METRICS] = metrics